requirements.txt
azure-functions==1.17.0
python-decouple==3.8
fastapi==0.109.0
uvicorn==0.27.0
pytest==7.4.4